requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.58
colorama==0.4.4
cssutils==2.3.0
decorator==5.0.9
demjson==2.2.4
Flask==1.1.2
gallery-dl==1.16.0
html5lib==1.1
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
LEPL==5.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mechanize==0.4.5
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.10.0
pycryptodome==3.10.1
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
retry==0.9.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tabulate==0.8.7
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
yoyo-migrations==7.3.2
dateparser==1.0.0
